nekidaem.django/requirements.txt
asgiref==3.3.1
Django==3.1.4
django-crispy-forms==1.10.0
psycopg2-binary==2.8.6
python-dotenv==0.15.0
pytz==2020.4
sqlparse==0.4.1
